Split consern between components CheckoutPage
Checkout Page renders Order where all
calculations done then result is passed
over to OrderSummary as props for rendering it
